feat: allow using route object in setting dynamic value + script
