+++ frontend/src/components/AppComponent.vue
@@ -84,6 +84,7 @@ const evaluationContext = computed(() => {
 		...store.resources,
 		...repeaterContext,
 		...componentContext?.value,
+		route: store.routeObject,
 	}
 })
